+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Yii2 EAuth Change Log
 =====================
 
 ### dev (13.01.2016)
-* Move response parsing from oauth to base service (fix #71)
+* Move response parsing from oauth to base service (#71)
+* Get user profile from Steam API (#70)
 
 ### 2.4.0 (03.01.2016)
 * Fixed error param names for Facebook (#63)